AI_Powered_Carbon_Capture/ai-engine/requirements.txt
# Core AI/ML Dependencies
tensorflow==2.13.0
scikit-learn==1.3.0
xgboost==1.7.6
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2

# Web Framework
fastapi==0.100.0
uvicorn[standard]==0.23.2
pydantic==2.0.3

# Async Processing
asyncio-mqtt==0.11.1

# Data Processing
joblib==1.3.2
scipy==1.11.1

# API and HTTP
requests==2.31.0
aiohttp==3.8.5

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and Monitoring
structlog==23.1.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.7.0
isort==5.12.0
flake8==6.0.0
mypy==1.5.1

# Optional: GPU Support (uncomment if using GPU)
# tensorflow-gpu==2.13.0

# Optional: Jupyter for notebooks
# jupyter==1.0.0
# jupyterlab==4.0.0